requirements.txt
discord[voice]
psycopg2-binary
types-psycopg2
lark
aiohttp
xmltodict
dicttoxml
gspread
python-dotenv
ruff
mypy
flask
waitress
types-waitress
packages/misc_cog